Fix compilation

+++ mistralrs-paged-attn/src/cuda/pagedattention.cu
@@ -772,7 +772,7 @@ extern "C" void paged_attention_v1(
     if (dtype == 2) {
       CALL_V1_LAUNCHER_BLOCK_SIZE(float, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     } else if (dtype == 0) {
-      CALL_V1_LAUNCHER_BLOCK_SIZE(half, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
+      CALL_V1_LAUNCHER_BLOCK_SIZE(uint16_t, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     } else if (dtype == 1) {
       CALL_V1_LAUNCHER_BLOCK_SIZE(__nv_bfloat16, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     }
@@ -781,7 +781,7 @@ extern "C" void paged_attention_v1(
     if (dtype == 2) {
       CALL_V1_LAUNCHER_BLOCK_SIZE(float, float, vllm::Fp8KVCacheDataType::kAuto);
     } else if (dtype == 0) {
-      CALL_V1_LAUNCHER_BLOCK_SIZE(half, half, vllm::Fp8KVCacheDataType::kAuto);
+      CALL_V1_LAUNCHER_BLOCK_SIZE(uint16_t, uint16_t, vllm::Fp8KVCacheDataType::kAuto);
     } else if (dtype == 1) {
       CALL_V1_LAUNCHER_BLOCK_SIZE(__nv_bfloat16, __nv_bfloat16, vllm::Fp8KVCacheDataType::kAuto);
     }
@@ -920,7 +920,7 @@ extern "C" void paged_attention_v2(
     if (dtype == 2) {
       CALL_V2_LAUNCHER_BLOCK_SIZE(float, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     } else if (dtype == 0) {
-      CALL_V2_LAUNCHER_BLOCK_SIZE(half, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
+      CALL_V2_LAUNCHER_BLOCK_SIZE(uint16_t, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     } else if (dtype == 1) {
       CALL_V2_LAUNCHER_BLOCK_SIZE(__nv_bfloat16, uint8_t, vllm::Fp8KVCacheDataType::kFp8E4M3);
     }
@@ -929,7 +929,7 @@ extern "C" void paged_attention_v2(
     if (dtype == 2) {
       CALL_V2_LAUNCHER_BLOCK_SIZE(float, float, vllm::Fp8KVCacheDataType::kAuto);
     } else if (dtype == 0) {
-      CALL_V2_LAUNCHER_BLOCK_SIZE(half, half, vllm::Fp8KVCacheDataType::kAuto);
+      CALL_V2_LAUNCHER_BLOCK_SIZE(uint16_t, uint16_t, vllm::Fp8KVCacheDataType::kAuto);
     } else if (dtype == 1) {
       CALL_V2_LAUNCHER_BLOCK_SIZE(__nv_bfloat16, __nv_bfloat16, vllm::Fp8KVCacheDataType::kAuto);
     }
